requirements.txt
streamlit
pandas
requests
plotly.express
